Move executor tests to nemo-runtime-executor
